Remove build, install and associated Build-Depends of Unihan database, support moved from source distribution to plugins
The build-stamp target and handling is removed, binary-arch and
binary-indep do not depend on build anymore and build is an empty target
now.  Build-Depends on unicode-data and bzip2 are no longer necessary.

The copyright notice for the previously included unihan.db is no longer
necessary as a consequence.  Fix copyright notice for Anki itself to
have years of copyright.

Closes: 551235