Transition to simpler patch management
Previously I used one branch per patch and merged upstream into those
then merged the branches to master and created / refreshed the patches
by diffing them to against upstream.

This is just too clever (for certain values of clever) for my own good
and is rather cumbersome actually.

With this commit all Debian patches are unapplied in the repository and
in the future the patches are updated with quilt and committed as simple
files as is usually done for simple packages.

Git-Dch: Ignore